Update from 2.0.0 release

+++ protected/console/controllers/DataController.php
@@ -2,15 +2,13 @@
 namespace console\controllers;
 
 use yii\console\Controller;
-use usni\library\components\UiDataManager;
 use usni\UsniAdaptor;
-use usni\library\utils\FileUtil;
 use usni\library\modules\users\models\User;
 /**
  * Run datamanager and install the data
  *
  * The command would be as follows yii data/index {managerClassName}.
- * managerClassName should be passed with forward slash for example backend/managers/NotificationDataManager.
+ * managerClassName should be passed with forward slash for example common/managers/AuthDataManager.
  * In case you run for a data manager like UsersDataManager which has notification data, you need to first delete it manually.
  * 
  * @package console\controllers
@@ -23,31 +21,25 @@ class DataController extends Controller
      */
     public function actionIndex($managerClassName)
     {
-        $path               = UsniAdaptor::app()->getRuntimePath();
         $managerClassName   = str_replace('/', '\\', $managerClassName);
         $modelClassName     = $managerClassName::getModelClassName();
-        $unserializedData   = UiDataManager::getUnserializedData('installdefaultdata.bin');
-        $tableName          =  $modelClassName::tableName();
-        $key                = array_search($tableName . '-' . $managerClassName, $unserializedData);
-        if($key != false)
+        //Delete existing records
+        if($modelClassName != null)
         {
-            unset($unserializedData[$key]);
-        }
-        $unserializedData = array_values($unserializedData);
-        FileUtil::writeFile($path, 'installdefaultdata.bin', 'w+', serialize($unserializedData));
-        $records        = $modelClassName::find()->all();
-        foreach($records as $record)
-        {
-            if($modelClassName == User::className())
+            $records            = $modelClassName::find()->all();
+            foreach($records as $record)
             {
-                if($record->id == User::SUPER_USER_ID || $record->username == 'storeowner')
+                if($modelClassName == User::className())
                 {
-                    continue;
+                    if($record->id == User::SUPER_USER_ID || $record->username == 'storeowner')
+                    {
+                        continue;
+                    }
                 }
+                $record->delete();
             }
-            $record->delete();
         }
-        $managerClassName::loadDefaultData();
-        $managerClassName::loadDemoData();
+        $managerClassName::getInstance()->loadDefaultData();
+        $managerClassName::getInstance()->loadDemoData();
     }
 }
